+++ tests/test_slack.py
@@ -7,7 +7,7 @@ import urllib.request
 from collections import deque
 from unittest.mock import patch
 
-from messages.slack import SlackWebHook
+from messages.slack import SlackWebhook
 from messages.eventloop import MESSAGELOOP
 
 
@@ -17,17 +17,17 @@ from messages.eventloop import MESSAGELOOP
 
 @pytest.fixture()
 def get_slack():
-    """Return a valid SlackWebHook object."""
-    return SlackWebHook('test_url', 'message', ['url1', 'url2'])
+    """Return a valid SlackWebhook object."""
+    return SlackWebhook('test_url', 'message', ['url1', 'url2'])
 
 
 ##############################################################################
-# TESTS: SlackWebHook.__init__
+# TESTS: SlackWebhook.__init__
 ##############################################################################
 
 def test_slack_init(get_slack):
     """
-    GIVEN a need for a SlackWebHook object
+    GIVEN a need for a SlackWebhook object
     WHEN instantiated
     THEN assert it is properly created
     """
@@ -40,14 +40,14 @@ def test_slack_init(get_slack):
 
 
 ##############################################################################
-# TESTS: SlackWebHook.construct_message
+# TESTS: SlackWebhook.construct_message
 ##############################################################################
 
 @patch.object(urllib.request, 'Request')
-@patch.object(SlackWebHook, 'add_attachments')
+@patch.object(SlackWebhook, 'add_attachments')
 def test_slack_construct_message(add_mock, req_mock, get_slack):
     """
-    GIVEN a valid SlackWebHook object
+    GIVEN a valid SlackWebhook object
     WHEN construct_message() is called
     THEN assert the message is properly created
     """
@@ -59,12 +59,12 @@ def test_slack_construct_message(add_mock, req_mock, get_slack):
 
 
 ##############################################################################
-# TESTS: SlackWebHook.add_attachments
+# TESTS: SlackWebhook.add_attachments
 ##############################################################################
 
 def test_slack_add_attachments_list(get_slack):
     """
-    GIVEN a valid SlackWebHook object with self.attach_urls = list
+    GIVEN a valid SlackWebhook object with self.attach_urls = list
     WHEN add_attachments() is called
     THEN assert the urls are properly attached to the message
     """
@@ -77,7 +77,7 @@ def test_slack_add_attachments_list(get_slack):
 
 def test_slack_add_attachments_str(get_slack):
     """
-    GIVEN a valid SlackWebHook object with self.attach_urls = str
+    GIVEN a valid SlackWebhook object with self.attach_urls = str
     WHEN add_attachments() is called
     THEN assert the urls are properly attached to the message
     """
@@ -89,14 +89,14 @@ def test_slack_add_attachments_str(get_slack):
 
 
 ##############################################################################
-# TESTS: SlackWebHook.send
+# TESTS: SlackWebhook.send
 ##############################################################################
 
 @patch.object(urllib.request, 'urlopen')
-@patch.object(SlackWebHook, 'construct_message')
+@patch.object(SlackWebhook, 'construct_message')
 def test_slack_send(con_mock, url_mock, get_slack, capsys):
     """
-    GIVEN a valid SlackWebHook object
+    GIVEN a valid SlackWebhook object
     WHEN send() is called
     THEN assert the proper send sequence occurs
     """
@@ -111,13 +111,13 @@ def test_slack_send(con_mock, url_mock, get_slack, capsys):
 
 
 ##############################################################################
-# TESTS: SlackWebHook.send_async
+# TESTS: SlackWebhook.send_async
 ##############################################################################
 
 @patch.object(MESSAGELOOP, 'add_message')
 def test_slack_send_async(msgloop_mock, get_slack):
     """
-    GIVEN a valid SlackWebHook object
+    GIVEN a valid SlackWebhook object
     WHEN send_async() is called
     THEN assert the proper send sequence occurs
     """